Minor fixes related to unsighned handling

+++ src/main/java/net/novaware/chip8/core/cpu/instruction/InstructionRegistry.java
@@ -8,6 +8,7 @@ import javax.inject.Inject;
 import java.util.*;
 
 import static net.novaware.chip8.core.cpu.instruction.InstructionType.*;
+import static net.novaware.chip8.core.util.UnsignedUtil.uint;
 import static net.novaware.chip8.core.util.UnsignedUtil.ushort;
 
 public class InstructionRegistry {
@@ -76,7 +77,7 @@ public class InstructionRegistry {
             short mask = masks.get(m).value();
 
             @Unsigned
-            short opcode = ushort(mask & instruction);
+            short opcode = ushort(uint(mask) & uint(instruction));
 
             var instrPerMask = defsIndex.get(mask);
             if (instrPerMask == null) {
